backend/requirements.txt
Django==2.1.1
django-cors-headers==2.4.0
django-environ==0.4.5
django-filter==2.0.0
djangorestframework==3.8.2
mysqlclient==1.3.13
pkg-resources==0.0.0
pytz==2018.5
